Minor Bug Fix and ViewModel Layer Implementation

+++ app/src/main/java/com/debanshu777/stocx/dataSource/model/Stock.kt
@@ -12,7 +12,7 @@ data class Stock(
     val sid: String,
 
     @SerializedName("change")
-    val change: Int,
+    val change: Double,
 
     @SerializedName("close")
     val close: Double,
